telegram_bot/requirements.txt
aiogram==2.14.3
aiohttp==3.7.4.post0
appdirs==1.4.4
async-timeout==3.0.1
attrs==21.2.0
Babel==2.9.1
black==21.7b0
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.1
idna==3.2
multidict==5.1.0
mypy-extensions==0.4.3
pathspec==0.9.0
python-decouple==3.4
pytz==2021.1
PyYAML==5.4.1
regex==2021.7.6
requests==2.26.0
six==1.16.0
tomli==1.0.4
typing-extensions==3.10.0.0
urllib3==1.26.6
yarl==1.6.3
